Split chart rendering and data into separate files
